solver: Eliminate unknown from consideration while solving
If we want a `Root` command to be deterministic irrespective of
previous values of variables we are solving for, we need to eliminate
them from consideration when figuring out if an equation is
well-defined. This ensures that the well-definedness is independent
from the possible pre-existence of variables that are in the list of
variables to be solved for.

This was illustrated for the `Driven Damped Oscillations` equation,
which would solve correctly if `γ=4.5 r/s` was set prior to solving.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>