array: Avoid slight risk of corruption in `det` commadn
The `det` code was pushing `this` on the stack repeatedly to make
room. This is technically incorrect, since `rt.push()` may cause a
garbage collection, invalidating `this`.

Push a garbage-collected pointer instead. Note that the value is not
really used afterwards, so memory corruption would only be possible if
the `this` pointer pushed caused a problem in a subsequent garbage
collection, which I believe would only occur if two garbage
collections occurred in a row while pushing the same `this` pointer.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>