lists: Fix subtle parsing bug at end of text
When the last parsed item happens to be before a byte that was an UTF8
marker, using `utf8_next` too naively would cause us to indicate a
parsed size that was beyond the end of text. This could result in
input validation error for `Text→Object`.

This was exposed by an example tests: `"{ 1 2 3 }" Text→Object`.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>